refactor(home): extract SectionHeader into its own file
Follows #4837. That PR extracted NativeTilesSection into
native-tiles-section.tsx but left SectionHeader in add-tile-drawer.tsx,
creating a two-file circular import (native-tiles-section imports
SectionHeader from add-tile-drawer, which imports NativeTilesSection back
from native-tiles-section). Moving SectionHeader to its own file removes
the cycle.
